+++ mathesar_ui/src/pages/schema/CreateExplorationTutorial.svelte
