+++ mathesar_ui/src/pages/schema/CreateTableButton.svelte
