+++ mathesar_ui/src/pages/schema/CreateTableTutorial.svelte
